Refactor object property access in 04_objects.js
This commit refactors the code in 04_objects.js to clarify the difference between accessing object properties using dot notation and bracket notation. It also fixes a bug where using bracket notation with an undefined variable would result in an error. Additionally, it introduces an example of using a variable as a key in an object.